mml/requirements.txt
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
docopt==0.6.2
filelock==3.9.0
fonttools==4.39.3
gdown==4.7.1
huggingface-hub==0.13.3
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
packaging==23.0
Pillow==9.3.0
pipreqs==0.4.11
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==6.0
regex==2023.3.23
requests==2.28.1
six==1.16.0
soupsieve==2.4
sympy==1.11.1
tokenizers==0.13.2
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118
tqdm==4.65.0
transformers==4.27.4
typing_extensions==4.4.0
urllib3==1.26.13
yarg==0.1.9

